feat: simulate start

+++ contracts/src/models/map.cairo
@@ -30,6 +30,6 @@ pub struct Bullet {
     #[key]
     pub bullet_id: u32,
     pub coords: Vec2,
-    pub speed: u32, // pixels per step
-    pub direction: u32, // in degrees
+    pub speed: i64, // pixels per step
+    pub direction: i64, // in degrees
 }
